Improved fix for issue #39.
Sizing numpy.memmap objects is handled by taking a configurable percentage of the mapped file size.  Mapped files are never or rarely completely mapped into memory.  The memory usage of a numpy.memmap object is typically only a small portion of the file size.  That portion depends on the file, the O/S and perhaps other system parameters.  A new asizeof function called amapped can be use to set that portion as a percentage (and get the previously set percentage).