Improved fix for issue #39.
Sizing numpy.memmap objects is handled by taking a configurable percentage of the mapped file size. Mapped files are rarely completely mapped into memory. The memory usage of a numpy.memmap object (and any mmap-ed file) is typically only a small portion of the file size. That portion depends on the file, the O/S and perhaps other system parameters.

A new asizeof function called amapped can be use to set that portion as a percentage (and get the previously set percentage). In addition, the size will be rounded up to a multiple of mmap.PAGESIZE, the O/S dependent virtual memory page size.