Updated asizeof.py for 3.13
The size of instances and dicts in Python 3.13+ respectively 3.6+ is much smaller, up to 5 times or more.  This asizeof version 25.07.17 has been updated to correctly size those objects.

Other changes are updates to the basic sizes, a check for running with/out the GIL enabled and mostly cosmetic ones.

NOTE, the test_asizeof.py module must be updated as well.